InventoryManager['lists'] add and remove listeners
Updated the recursive container crawling function to add a
'child_removed' listener that removes the list from the
InventoryManager['lists'] array when a grocery list is removed.

Fixed the logic in recurseContainers() to add the child_added listener
to all children, whether or not they are grocery lists (allows grocery
lists nested under non-grocery lists to be seen).